Genesis day 3
And God said, 'Let the users be gathered to one place, and let items appear.' ...
Then God said, 'Let the items have labels: words generated by faker.' And it was so. The users produced comments: faker generated paragrahps, associated with an item. And God saw that it was good. And there were relations, and there was seeded data--the third day.